+++ src/Services/axiosInstance.tsx
@@ -5,7 +5,7 @@ const BASE_URLS: Record<number, string> = {
   5001: "http://localhost:5001/",
   5002: "http://localhost:5002/",
   5003: "http://localhost:5003/",
-  5004: "http://localhost:5004/",
+  5004: "billbizz-alb-front-269260655.ap-south-1.elb.amazonaws.com:5004",
   5005: "http://localhost:5005/",
   5006: "http://localhost:5006/",
   5007: "http://localhost:5007/",
