+++ src/layout/SubHeader/SubHeader.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from "react";
 import navlist from "../../assets/constants";
 import { Link } from "react-router-dom";
+import ItemEllipsis from "../../Components/ellipsis/Ellipsis";
 
 type Props = {
   activeIndex: number | null;
@@ -22,25 +23,26 @@ const SubHeader = ({ activeIndex }: Props) => {
   };
 
   return (
-    <div className="bg-BgSubhead flex mx-7 pl-7 py-5 my-4 rounded-full items-center gap-4">
-      <div className="bg-white px-4 py-2 rounded-full text-sm">
-         Home
+    <div className="bg-BgSubhead flex mx-7 justify-between px-7 py-5 my-4 items-center rounded-full ">
+      <div className="flex items-center gap-4">
+        <div className="bg-white px-4 py-2 rounded-full text-sm">Home</div>
+        {activeIndex !== null &&
+          navlist[activeIndex] &&
+          navlist[activeIndex].subhead &&
+          navlist[activeIndex].subhead.map((item, index) => (
+            <Link to={item.subRoute} key={index}>
+              <div
+                className={`font-medium py-2 px-4 rounded-full cursor-pointer ${
+                  selectedIndex === index ? "bg-white" : "hover:bg-white"
+                }`}
+                onClick={() => handleSelect(index)}
+              >
+                {item.headName}
+              </div>
+            </Link>
+          ))}
       </div>
-      {activeIndex !== null &&
-        navlist[activeIndex] &&
-        navlist[activeIndex].subhead &&
-        navlist[activeIndex].subhead.map((item, index) => (
-          <Link to={item.subRoute}>
-            <div
-              className={`font-medium py-2 px-4 rounded-full cursor-pointer ${
-                selectedIndex === index ? "bg-white" : "hover:bg-white"
-              }`}
-              onClick={() => handleSelect(index)}
-            >
-              {item.headName}
-            </div>
-          </Link>
-        ))}
+      <div>{activeIndex === 1 && <ItemEllipsis />}</div>
     </div>
   );
 };
